Rename XWizardMigrate to XWizardMigrateBluesky in anticipation of supporting Mastodon too
